Remove blending during RGBA painting
painting in RGB/RGBA mode is now much more... 'pixely'.
Instead of trying to use the brush alpha channel to blend on to the
image, now we just replace pixels.
Different to most paint programs, but more suited to pixel-editing.

Definitely a need for a proper alpha-blending draw mode, but that
can wait a bit (and requires some plumbing - eg we currently assume
tools can overdraw as much as we want. This simplifies the drawing
code, but breaks down completely during blending)